requirements.txt
aerich==0.6.2
aiosqlite==0.17.0
anyio==3.5.0
asyncpg==0.25.0
attrs==21.4.0
Automat==20.2.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.2
ddlparse==1.10.0
dependency-injector==4.39.1
dictdiffer==0.9.0
dnspython==2.2.1
email-validator==1.1.3
environs==9.5.0
fastapi==0.75.0
fastapi-users==9.3.0
fastapi-users-db-tortoise==2.0.0
greenlet==1.1.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
iso8601==0.1.16
makefun==1.13.1
marshmallow==3.15.0
packaging==21.3
passlib==1.7.4
priority==2.0.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
pypika-tortoise==0.1.4
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
toml==0.10.2
tomlkit==0.10.1
tortoise-orm==0.18.1
Twisted==22.2.0
twisted-iocpsupport==1.0.2
typing_extensions==4.1.1
wsproto==1.1.0
zipp==3.7.0
zope.interface==5.4.0
